pyproject.toml
[project]
name = "vibe-2025"
version = "0.1.0"
description = "AI-powered MCQ generator — FastAPI backend and React frontend monorepo template"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langsmith>=0.1.107",
    "python-dotenv>=1.0.1",
    "pypdf>=4.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "boto3>=1.40.59",
    "botocore>=1.40.59",
    "sqlalchemy>=2.0.36",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py313"
src = ["fastapi"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
